Move snapshot to its own file
It's about to gain non-trivial functionality.
